requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-neo4j
qdrant-client
neo4j
python-dotenv
pydantic
